requirements.txt
Flask>=2.0
open-interpreter>=0.1.8
fastapi
uvicorn
websockets
jinja2
python-multipart 